LoginVM 이동
애플 로그인 뷰 모델에 이동 작업
