+++ Run-It/LoginView/LoginVM.swift
@@ -9,6 +9,8 @@ import Foundation
 import FirebaseAuth
 import KakaoSDKAuth
 import KakaoSDKUser
+import AuthenticationServices
+import CryptoKit
 
 
 // MARK: - Firebase 로그인
@@ -153,6 +155,98 @@ func bringKakaoInfo()
     }
 }
 
+//MARK: - 애플 로그인
+class LoginVM: NSObject, ASAuthorizationControllerDelegate, ASAuthorizationControllerPresentationContextProviding
+{
+    
+    var currentNonce: String?
+    var presentationAnchor: ASPresentationAnchor?
+    
+    func setPresentationAnchor(_ anchor: ASPresentationAnchor) 
+    {
+        presentationAnchor = anchor
+    }
+
+    func appleLogin() 
+    {
+        let nonce = randomNonceString()
+        currentNonce = nonce
+        let appleIDProvider = ASAuthorizationAppleIDProvider()
+        let request = appleIDProvider.createRequest()
+        request.requestedScopes = [.fullName, .email]
+        request.nonce = sha256(nonce)
+        
+        let authorizationController = ASAuthorizationController(authorizationRequests: [request])
+        authorizationController.delegate = self
+        if presentationAnchor != nil
+        {
+            authorizationController.presentationContextProvider = self
+            authorizationController.performRequests()
+        }
+        
+        else
+        {
+            fatalError("presentationAnchor 가 설정되지 않음")
+        }
+        sendTask(task: "successLogin")
+    }
+    
+    func sha256(_ input: String) -> String
+    {
+        let inputData = Data(input.utf8)
+        let hashedData = SHA256.hash(data: inputData)
+        let hashString = hashedData.compactMap {
+            return String(format: "%02x", $0)
+        }.joined()
+        
+        return hashString
+    }
+    
+    func randomNonceString(length: Int = 32) -> String
+    {
+        precondition(length > 0)
+        let charset: Array<Character> =
+        Array("0123456789ABCDEFGHIJKLMNOPQRSTUVXYZabcdefghijklmnopqrstuvwxyz-._")
+        var result = ""
+        var remainingLength = length
+        
+        while remainingLength > 0 {
+            let randoms: [UInt8] = (0 ..< 16).map
+            {   _ in
+                var random: UInt8 = 0
+                let errorCode = SecRandomCopyBytes(kSecRandomDefault, 1, &random)
+                if errorCode != errSecSuccess
+                {
+                    fatalError("nonce를 생성 실패 - \(errorCode)")
+                }
+                return random
+            }
+            
+            randoms.forEach
+            {   random in
+                if remainingLength == 0
+                {
+                    return
+                }
+                
+                if random < charset.count
+                {
+                    result.append(charset[Int(random)])
+                    remainingLength -= 1
+                }
+            }
+        }
+        
+        return result
+    }
+    
+    func presentationAnchor(for controller: ASAuthorizationController) -> ASPresentationAnchor
+    {
+        guard let anchor = presentationAnchor else { fatalError("presentationAnchor 가 설정되지 않음") }
+        
+        return anchor
+    }
+}
 
 // MARK: - Notification
 func sendTask(task: String)
