+++ Run-It/LoginView/LoginViewController.swift
@@ -15,13 +15,14 @@ import KakaoSDKUser
 import AuthenticationServices
 import CryptoKit
 
-var currentNonce: String?
 
 class LoginViewController: UIViewController
 {
     var persistentContainer: NSPersistentContainer? {
         (UIApplication.shared.delegate as? AppDelegate)?.persistentContainer
     }
+    
+    let loginVM = LoginVM()
     
     let loginLogo = UIImageView(image: UIImage(named: "LoginLogo"))
     
@@ -371,12 +372,9 @@ class LoginViewController: UIViewController
     
     @objc func touchedAppleLoginButton()
     {
-        appleLogin()
+        loginVM.setPresentationAnchor(self.view.window!)
+        loginVM.appleLogin()
         self.checkData(loginType: "Apple", email: "apple")
-        let VC = MainTabBarViewController()
-        
-        VC.modalPresentationStyle = .fullScreen
-        self.present(VC, animated: true, completion: nil)
     }
     
     @objc func touchedFindEmailButton()
@@ -507,103 +505,4 @@ extension LoginViewController: UITextFieldDelegate
         passwordTextField.becomeFirstResponder()
     }
 
-}
-
-extension LoginViewController: ASAuthorizationControllerDelegate
-{
-    func authorizationController(controller: ASAuthorizationController, didCompleteWithAuthorization authorization: ASAuthorization) 
-    {
-        if let appleIDCredential = authorization.credential as? ASAuthorizationAppleIDCredential
-        {
-            guard let nonce = currentNonce  else { fatalError("로그인 요청이 전송되지 않았습니다") }
-            
-            guard let appleIDToken = appleIDCredential.identityToken    else { print("토큰을 가져올 수 없습니다"); return }
-            
-            guard let idTokenString = String(data: appleIDToken, encoding: .utf8)   else { print("토큰 문자열을 직렬화 할 수 없습니다"); return }
-            
-            let credential = OAuthProvider.credential(withProviderID: "apple.com", idToken: idTokenString, rawNonce: nonce)
-            
-            Auth.auth().signIn(with: credential)
-            {   authResult, error in
-                if let error = error
-                {
-                    print("애플 로그인 오류: \(error)")
-                    return
-                }
-                
-            }
-        }
-    }
-}
-
-extension LoginViewController: ASAuthorizationControllerPresentationContextProviding
-{
-    func appleLogin()
-    {
-        let nonce = randomNonceString()
-        currentNonce = nonce
-        let appleIDProvider = ASAuthorizationAppleIDProvider()
-        let request = appleIDProvider.createRequest()
-        request.requestedScopes = [.fullName, .email]
-        request.nonce = sha256(nonce)
-        
-        let authorizationController = ASAuthorizationController(authorizationRequests: [request])
-        authorizationController.delegate = self
-        authorizationController.presentationContextProvider = self
-        authorizationController.performRequests()
-    }
-    
-    func sha256(_ input: String) -> String
-    {
-        let inputData = Data(input.utf8)
-        let hashedData = SHA256.hash(data: inputData)
-        let hashString = hashedData.compactMap {
-            return String(format: "%02x", $0)
-        }.joined()
-        
-        return hashString
-    }
-    
-    func randomNonceString(length: Int = 32) -> String
-    {
-        precondition(length > 0)
-        let charset: Array<Character> =
-        Array("0123456789ABCDEFGHIJKLMNOPQRSTUVXYZabcdefghijklmnopqrstuvwxyz-._")
-        var result = ""
-        var remainingLength = length
-        
-        while remainingLength > 0 {
-            let randoms: [UInt8] = (0 ..< 16).map
-            { _ in
-                var random: UInt8 = 0
-                let errorCode = SecRandomCopyBytes(kSecRandomDefault, 1, &random)
-                if errorCode != errSecSuccess
-                {
-                    fatalError("nonce를 생성 실패 - \(errorCode)")
-                }
-                return random
-            }
-            
-            randoms.forEach
-            {   random in
-                if remainingLength == 0
-                {
-                    return
-                }
-                
-                if random < charset.count
-                {
-                    result.append(charset[Int(random)])
-                    remainingLength -= 1
-                }
-            }
-        }
-        
-        return result
-    }
-    
-    func presentationAnchor(for controller: ASAuthorizationController) -> ASPresentationAnchor
-    {
-        return self.view.window!
-    }
 }
